refactor(color-picker): components
