+++ packages/components/color-picker/components/format/config.ts
