+++ packages/components/color-picker/components/format/style/index.js
