+++ packages/components/color-picker/components/panel/header.tsx
@@ -1,11 +1,11 @@
 import { defineComponent, PropType, ref, watch } from 'vue';
 
-import props from '../props';
+import props from '../../props';
 import { COLOR_MODES } from '@tdesign/common-js/color-picker/constants';
-import { RadioGroup as TRadioGroup, RadioButton as TRadioButton } from '../../radio';
-import { TdColorModes } from '../types';
-import { useBaseClassName } from '../hooks';
-import { useConfig } from '../../hooks';
+import { RadioGroup as TRadioGroup, RadioButton as TRadioButton } from '../../../radio';
+import { TdColorModes } from '../../types';
+import { useBaseClassName } from '../../hooks';
+import { useConfig } from '../../../hooks';
 
 export default defineComponent({
   name: 'PanelHeader',
