+++ packages/components/color-picker/components/panel/hue.tsx
@@ -1,6 +1,6 @@
 import { defineComponent } from 'vue';
 import ColorSlider from './slider';
-import { useBaseClassName } from '../hooks';
+import { useBaseClassName } from '../../hooks';
 import baseProps from './base-props';
 
 export default defineComponent({
