+++ packages/components/color-picker/components/panel/linear-gradient.tsx
@@ -1,10 +1,10 @@
 import { defineComponent, onBeforeUnmount, onMounted, reactive, ref, watch } from 'vue';
 import { cloneDeep } from 'lodash-es';
 import { GRADIENT_SLIDER_DEFAULT_WIDTH } from '@tdesign/common-js/color-picker/constants';
-import { genGradientPoint, gradientColors2string, GradientColorPoint } from '../utils';
-import { InputNumber as TInputNumber } from '../../input-number';
-import { useBaseClassName } from '../hooks';
-import { useCommonClassName } from '../../hooks/useConfig';
+import { genGradientPoint, gradientColors2string, GradientColorPoint } from '../../utils';
+import { InputNumber as TInputNumber } from '../../../input-number';
+import { useBaseClassName } from '../../hooks';
+import { useCommonClassName } from '../../../hooks/useConfig';
 import baseProps from './base-props';
 
 const DELETE_KEYS: string[] = ['delete', 'backspace'];
