+++ packages/components/color-picker/components/panel/slider.tsx
@@ -1,7 +1,7 @@
 import { computed, defineComponent, onBeforeUnmount, onMounted, PropType, reactive, ref } from 'vue';
 import { SLIDER_DEFAULT_WIDTH } from '@tdesign/common-js/color-picker/constants';
-import { Draggable, Coordinate } from '../utils';
-import { useBaseClassName } from '../hooks';
+import { Draggable, Coordinate } from '../../utils';
+import { useBaseClassName } from '../../hooks';
 import baseProps from './base-props';
 
 export default defineComponent({
