+++ packages/components/color-picker/components/panel/style/index.js
