+++ packages/components/color-picker/components/trigger/style/css.js
@@ -1,0 +1,1 @@
+import '../../../style/index.css';
