+++ packages/components/color-picker/components/trigger/style/index.js
@@ -1,0 +1,1 @@
+import '../../../style/index.js';
