--- packages/components/color-picker/panel/style/css.js
@@ -1,1 +1,0 @@
-import '../../style/index.css';
