--- packages/components/color-picker/panel/style/index.js
@@ -1,1 +1,0 @@
-import '../../style/index.js';
